src/requirements.txt
beautifulsoup4==4.12.2
colorama==0.4.4
dnspython==2.4.2
pyasn==1.6.2
python_nmap==0.7.1
Requests==2.31.0
